[DEAF-300] refactor export to file dialog and add en translations
