+++ app/i18n/locales/en-US.json
@@ -37,6 +37,8 @@
   "global.language.korean": "Korean",
   "global.passwordInstructions": "Note that password needs to be at least 7 characters long, and have at least 1 uppercase, 1 lowercase letter and 1 number.",
   "global.unit.ada": "Ada",
+  "global.spendingPasswordLabel": "Spending Password",
+  "global.spendingPasswordPlaceholder": "Password",
   "inline.editing.dropdown.changesSaved": "Your changes have been saved",
   "inline.editing.input.cancel.label": "cancel",
   "inline.editing.input.change.label": "change",
@@ -103,10 +105,9 @@
   "wallet.create.dialog.title": "Create wallet",
   "wallet.create.dialog.walletNameHint": "e.g: Shopping Wallet",
   "wallet.create.dialog.walletPasswordLabel": "Wallet password",
-  "wallet.export.choices.tab.title.full": "Full",
-  "wallet.export.choices.tab.title.paperWallet": "Paper wallet",
-  "wallet.export.choices.tab.title.readOnly": "Read-only",
-  "wallet.export.dialog.headline": "Export Wallet",
+  "wallet.exportToFile.dialog.headline": "Export Wallet",
+  "wallet.exportToFile.dialog.introduction": "You are exporting <strong>{walletName}</strong> to a file.",
+  "wallet.exportToFile.dialog.submit.label": "Export",
   "wallet.key.import.dialog.headline": "Import Wallet",
   "wallet.key.import.dialog.keyFileHint": "Drop file here or click to choose",
   "wallet.key.import.dialog.keyFileLabel": "Upload your key file",
